+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 * [#2576](https://github.com/xmake-io/xmake/issues/2576): More flexible package fetching from cmake
 * [#2577](https://github.com/xmake-io/xmake/issues/2577): Improve add_headerfiles(), add `{install = false}` support
+* [#2603](https://github.com/xmake-io/xmake/issues/2603): Disable `-fdirectives-only` for ccache by default
 
 ## v2.6.9
 
@@ -1355,6 +1356,7 @@
 
 * [#2576](https://github.com/xmake-io/xmake/issues/2576): More flexible package fetching from cmake
 * [#2577](https://github.com/xmake-io/xmake/issues/2577): 改进 add_headerfiles()，增加 `{install = false}` 支持
+* [#2603](https://github.com/xmake-io/xmake/issues/2603): 为 ccache 默认禁用 `-fdirectives-only`
 
 ## v2.6.9